index collision boxes with a grid instead of rtree
The combined running time of placeCollisionFeature and
insertCollisionFeature is now 2/3rds of what it used to be.

The grid index divids a square into n x n cells. When a bbox is
inserted, it is added to the array of every cell it intersects. When
querying, look in all the cells that intersect the query box and then
compare all the individual bboxes contained in that cell against the
query box.

This kind of index is faster in this specific case because it's
characteristics better match the work we're using it for.

Queries can take 1.3x as long (slower) with the grid index.
But insertions often take < 0.1x as long with the grid index.
Our collision detection code does a lot of insertions so this is a
worthwhile tradeoff.

The grid index also takes advantage of the fact that label boxes are
fairly evenly sized and evenly distributed.

The grid index can also be serialized to an ArrayBuffer. The ArrayBuffer
version can be queried without deserializing all the individual objects
first. This is not currently used.