use grid index instead of rtree for querying features
Why?
Both indexes have comparable querying performance and Grid is
transferrable. Insertion is faster with Grid than rbush but that's not
the main benefit.

Point queries are really fast for both. Each is sometimes faster than
the other. Therre is no clear winner.

Large, full-screen bbox queries are 10% slower with Grid than rbush but
difference is not significant (0.2% of FeatureTree.query).

comparisons were done with mapbox-streets-v6